Move recording tests to its own namespace in Echo.IntegrationTests
